chore: merge files by purpose
