+++ runner/internal/metrics/aggregate.go
@@ -1,6 +1,7 @@
 package metrics
 
 import (
+	"strconv"
 	"time"
 
 	"github.com/benchttp/engine/internal/timestats"
@@ -63,4 +64,40 @@ func NewAggregate(records []recorder.Record) (agg Aggregate) {
 	agg.RequestEventTimes = computeRequestEventTimes(records)
 
 	return agg
+}
+
+// Special compute helpers.
+
+func computeRequestEventTimes(records []recorder.Record) map[string]timestats.TimeStats {
+	events := getFlatRelativeTimeEvents(records)
+
+	timesByEvent := map[string][]time.Duration{}
+
+	for _, e := range events {
+		timesByEvent[e.Name] = append(timesByEvent[e.Name], e.Time)
+	}
+
+	statsByEvent := map[string]timestats.TimeStats{}
+
+	for e, times := range timesByEvent {
+		statsByEvent[e] = timestats.New(times)
+	}
+
+	return statsByEvent
+}
+
+func getFlatRelativeTimeEvents(records []recorder.Record) []recorder.Event {
+	events := []recorder.Event{}
+	for _, record := range records {
+		events = append(events, record.RelativeTimeEvents()...)
+	}
+	return events
+}
+
+func computeStatusCodesDistribution(records []recorder.Record) map[string]int {
+	statuses := map[string]int{}
+	for _, rec := range records {
+		statuses[strconv.Itoa(rec.Code)]++
+	}
+	return statuses
 }
